Fix for template creation with GA

+++ apps/tables/signals.py
@@ -12,4 +12,8 @@ def delete_bigquery_table(sender, instance, *args, **kwargs):
     if settings.MOCK_REMOTE_OBJECT_DELETION:
         return
 
+    # hotfix for cloned templates where bq dataset name is invalid
+    if " copy " in instance.bq_id:
+        return
+
     bigquery_client().delete_table(instance.bq_id, not_found_ok=True)
